Dialogue + Relationship Manager - Bojinca Ioana-Lucia and Ion Corina-Maria.
The implemented feature is a 2D Dialogue and Relationship Manager.
The dialogue will be displayed in a UI panel element. The user goes through the dialogue by pressing the “D” key.

At certain points, the player can respond to the dialogue by clicking one of two choices.
These choices affect the relationship the player has with the ‘narrator’ character.  Each choice has a positive or negative effect on the relationship, based on the action chosen. The changes in the relationship can be viewed in the Slider (value and color changes) as well as in the Text element below the slider (will display “relationship points” and the current relationship - friendly, neutral, unfriendly or hostile).

The multiple choice dialogue was created using the Ink editor and Unity Package.

Dependecies: Ink (https://www.inklestudios.com/ink/)

Tasks:
Ion Corina - creating the dialogue logic, testing the demo
Bojinca Ioana  - created the relationship manager logic